requirements.txt
django>=5.2,<6.0
python-dotenv>=1.0
pillow>=10.0
google-genai>=1.0
requests>=2.31
